+++ src/renderer/pages.ts
@@ -1,0 +1,7 @@
+// an enum with the name and the route of the page
+export enum Pages {
+  SIGN_IN = '/sign-in',
+  HOME = '/home',
+  CHOOSE_PATH = '/',
+  GENERATE_WALLET = '/generate-wallet',
+}